Adding software architecture picture and
- Shorten phpMyAdmin container name to jbt_admin
- Change create script option from build to no-cache
- Using host.docker.internal:7006 for database
- Some small docu changes

+++ scripts/create.sh
@@ -1,13 +1,15 @@
 #!/bin/bash -e
 #
-# create.sh - delete all docker containers, build them new and install Joomla from git branches
+# create.sh - Delete all docker containers, build them new and install Joomla from the git branches.
+#   create.sh
+#   create.sh no-cache
 #
 # MIT License, Copyright (c) 2024 Heiko Lübbe
 # https://github.com/muhme/joomla-branches-tester
 
 source scripts/helper.sh
 
-# zeroth check host.docker.internal entry 
+# Zeroth check host.docker.internal entry
 HOSTS_FILE="/etc/hosts"
 EXPECTED_ENTRY="127.0.0.1[[:space:]]+host.docker.internal"
 if grep -Eq "${EXPECTED_ENTRY}" "$HOSTS_FILE"; then
@@ -20,13 +22,15 @@ fi
 # First delete all docker containters
 scripts/clean.sh
 
-for version in "${VERSIONS[@]}"
-do
-  log "Removing directory branch_${version}"
-  rm -rf "branch_${version}"
+# Clean up branch directories if necessary
+for version in "${VERSIONS[@]}"; do
+  if [ -d "branch_${version}" ]; then
+    log "Removing directory branch_${version}"
+    rm -rf "branch_${version}"
+  fi
 done
 
-if [ $# -eq 1 ] && [ "$1" = "build" ] ; then
+if [ $# -eq 1 ] && [ "$1" = "no-cache" ]; then
   log "Docker compose build --no-cache"
   docker compose build --no-cache
 fi
@@ -34,25 +38,24 @@ fi
 log "Docker compose up"
 docker compose up -d
 
-for version in "${VERSIONS[@]}"
-do
-  # if the copying has not yet been completed, then we have to wait, or we will get e.g.
-  # rm: cannot remove '/var/www/html/libraries/vendor': Directory not empty
+for version in "${VERSIONS[@]}"; do
+  # If the copying has not yet been completed, then we have to wait, or we will get e.g.
+  # rm: cannot remove '/var/www/html/libraries/vendor': Directory not empty.
   max_retries=120
-  for ((i=1; i<$max_retries; i++)); do
-  docker logs "jbt_${version}" 2>&1 | grep 'This server is now configured to run Joomla!' && break || {
+  for ((i = 1; i < $max_retries; i++)); do
+    docker logs "jbt_${version}" 2>&1 | grep 'This server is now configured to run Joomla!' && break || {
       log "Waiting for original Joomla installation, attempt ${i}/${max_retries}"
       sleep 1
     }
   done
   if [ $i -ge $max_retries ]; then
-      error "Failed after $max_retries attempts, giving up"
-      exit 1
+    error "Failed after $max_retries attempts, giving up"
+    exit 1
   fi
   log "jbt_${version} – Deleting orignal Joomla installation"
   docker exec -it "jbt_${version}" bash -c 'rm -rf /var/www/html/* && rm -rf /var/www/html/.??*'
 
-  # disable the disabled PHP error logging
+  # Move away the disabled PHP error logging.
   log "jbt_${version} – Show PHP warnings"
   docker exec -it "jbt_${version}" bash -c 'mv /usr/local/etc/php/conf.d/error-logging.ini /usr/local/etc/php/conf.d/error-logging.ini.DISABLED'
 
@@ -61,7 +64,7 @@ do
     apt-get upgrade -y && \
     curl -fsSL https://deb.nodesource.com/setup_22.x | bash - && \
     apt-get install -y git vim nodejs iputils-ping net-tools'
-  # aditional having vim, ping, netstat
+  # Aditional having vim, ping, netstat
 
   branch=$(branchName "${version}")
   log "jbt_${version} – cloning ${branch} branch into directory branch_${version}"
@@ -91,16 +94,20 @@ do
   # adopt e.g.:
   #   >     db_name: 'test_joomla_44'
   #   >     db_prefix: 'jos44_',
-  #   >     db_host: 'mysql',
+  #   >     db_host: 'host.docker.internal',
+  #   >     db_port: '7006',
   #   >     baseUrl: 'http://host.docker.internal:7044',
   #   >     db_password: 'root',
   #   >     smtp_host: 'host.docker.internal',
   #   >     smtp_port: '7025',
+
+  #     -e \"s/baseUrl: .*/baseUrl: 'http:\/\/jbt_${version}\/',/\" \
   docker exec -it "jbt_${version}" bash -c "cd /var/www/html && sed \
     -e \"s/db_name: .*/db_name: 'test_joomla_${version}',/\" \
     -e \"s/db_prefix: .*/db_prefix: 'jos${version}_',/\" \
-    -e \"s/db_host: .*/db_host: 'mysql',/\" \
-    -e \"s/baseUrl: .*/baseUrl: 'http:\/\/jbt_${version}\/',/\" \
+    -e \"s/db_host: .*/db_host: 'host.docker.internal',/\" \
+    -e \"s/db_port: .*/db_port: '7006',/\" \
+    -e \"s/baseUrl: .*/baseUrl: 'http:\/\/host.docker.internal:70${version}\/',/\" \
     -e \"s/db_password: .*/db_password: 'root',/\" \
     -e \"s/smtp_host: .*/smtp_host: 'host.docker.internal',/\" \
     -e \"s/smtp_port: .*/smtp_port: '7025',/\" \
@@ -122,3 +129,8 @@ do
   log "jbt_${version} – Enable MySQL user root login with password"
   docker exec -it jbt_mysql mysql -uroot -proot -e "ALTER USER 'root'@'%' IDENTIFIED WITH mysql_native_password BY 'root';"
 done
+
+log "Aditional having vim, ping and netstat in jbt_cypress container"
+docker exec -it jbt_cypress sh -c "apt-get update && apt-get install -y git vim iputils-ping net-tools"
+
+log "Work is done, you can now use scripts/tests.sh for System Tests or scripts/patchtester.sh to install Joomla Test Patcher"
